requirements.txt
google-genai>=1.0.0
pyaudio>=0.2.14
opencv-python>=4.8.0
python-dotenv>=1.0.0
